--- src/OSDP.Net/Tracing/DummyTracer.cs
@@ -1,10 +1,0 @@
-﻿namespace OSDP.Net.Tracing;
-
-internal class DummyTracer : ITracer
-{
-    public void Dispose() { }
-
-    public void Trace(TraceEntry trace) { }
-
-    public readonly static DummyTracer Default = new DummyTracer();
-}
